Think i'm along the right path now.
I believe that the octet held in memory (e.g. pedestrian light) repesents
a sequence of lights of the same type, in their on or off states.

This seems logical. After much debugging, it was found that HWIF.TRAFFIC_LIGHT
would grab the object in the body, without a declaration.